chore(quill-charts): split hover resolution and pill geometry out of draw-bar-chart
draw-bar-chart.ts had accumulated non-drawing code. Move it into dedicated utils so the file holds only canvas drawing:

- resolve-bar-hover.ts: resolveBarHoverItems + its types (pure hover resolution, mirrors resolve-clicked-bar-series.ts)
- stack-pills.ts: stackPillRects pill geometry, shared by the static drawer and the hover resolver

draw-bar-chart.ts now exports only drawBarChartStatic and drawBarHoverItems.

Generated-By: PostHog Code
Task-Id: b5369a5d-adbd-41e0-ae3e-31841364ce11
